Added traced on Modbus - Bug fix on Reports
Bug fix 1) Reports generated on startup.
The code takes the current day and changes it to the report day. If the
report day is after the current day (ex: current = wednesday and report
= friday) everything is ok. But if the report day is before the current
day in the week (ex: current = friday and report = tuesday) then the
code believes that we are already ready to generate a report and does
it.
I added a test to correct this and add 1 week to the report day.

Bug fix 2) Reports are generated two times instead of one.
I believe that the reports are a bit long to generate and happen to
check two times if the reports have to be generated. If it's the case
then the flags used are reseted and we do the job twice...
I improved a bit the logic I used based on GRAFCET to secure that.