Added LogRate and RefreshRate handling to digital MaxioIO class.
Now LogRate and RefreshRate should be handled as expected (was not
implemented yet).
If the value changes between two RefreshRate or if the time elapsed
exceeds the LogRate then the value is updated.
Otherwise we do nothing.

This allows the Vigie to know very quickly if a digital input has
changed.