Big commit - Reports, Logs, Workers, ModbusRequest, SignalType impacted.
ModbusRequest, SignalType, VigieAcquisition : Added a column "precision"
in the database to specify the level of detail the data should hold.

VigieReport : New option to specify if the reports are to be generated
daily, weekly or monthly (monthly not implemented yet as it would be
nice to check if the day of the month specified exists and do stuff if
not and so on ...).

Logs : Reduces the length of logs by cutting the classname out of it.

Workers : Changed from LinkedList to ConcurrentLinkedQueue because of
concurrent access problems. ConcurrentLinkedQueue is overkill but
simpler to implement than the Collections.synchronisedList(...) method.